Improve GoldenEye models and code organization
