+++ example/src/main/java/co/infinum/example/Extensions.kt
@@ -1,6 +1,9 @@
 package co.infinum.example
 
-import co.infinum.goldeneye.*
+import co.infinum.goldeneye.models.FlashMode
+import co.infinum.goldeneye.models.FocusMode
+import co.infinum.goldeneye.models.PreviewScale
+import co.infinum.goldeneye.models.Size
 
 fun FocusMode.convertToString() = name.toLowerCase()
 fun FlashMode.convertToString() = name.toLowerCase()
