+++ goldeneye/src/main/java/co/infinum/goldeneye/extensions/Any.kt
@@ -1,0 +1,21 @@
+package co.infinum.goldeneye.extensions
+
+import android.os.Handler
+import android.os.Looper
+import java.util.concurrent.Executors
+
+internal val backgroundExecutor = Executors.newSingleThreadExecutor()
+internal val mainHandler = Handler(Looper.getMainLooper())
+
+internal fun <T1, T2> ifNotNull(p1: T1?, p2: T2?, action: (T1, T2) -> Unit) {
+    if (p1 != null && p2 != null) {
+        action(p1, p2)
+    }
+}
+
+internal fun <T> async(task: () -> T?, onResult: (T?) -> Unit) {
+    backgroundExecutor.execute {
+        val result = task()
+        mainHandler.post { onResult(result) }
+    }
+}
