+++ goldeneye/src/main/java/co/infinum/goldeneye/extensions/ByteArray.kt
@@ -1,0 +1,10 @@
+package co.infinum.goldeneye.extensions
+
+import android.graphics.BitmapFactory
+
+internal fun ByteArray.toBitmap() =
+    try {
+        BitmapFactory.decodeByteArray(this, 0, size)
+    } catch (t: Throwable) {
+        null
+    }
