+++ goldeneye/src/main/java/co/infinum/goldeneye/extensions/View.kt
@@ -1,0 +1,6 @@
+package co.infinum.goldeneye.extensions
+
+import android.view.View
+
+internal fun View.isMeasured() = height > 0 && width > 0
+internal fun View.isNotMeasured() = isMeasured().not()
