+++ goldeneye/src/main/java/co/infinum/goldeneye/models/CameraInfo.kt
@@ -1,0 +1,7 @@
+package co.infinum.goldeneye.models
+
+data class CameraInfo(
+    val id: Int,
+    val orientation: Int,
+    val facing: Facing
+)
