+++ goldeneye/src/main/java/co/infinum/goldeneye/models/Facing.kt
@@ -1,0 +1,5 @@
+package co.infinum.goldeneye.models
+
+enum class Facing {
+    BACK, FRONT
+}
