+++ goldeneye/src/main/java/co/infinum/goldeneye/models/PreviewScale.kt
@@ -1,0 +1,5 @@
+package co.infinum.goldeneye.models
+
+enum class PreviewScale {
+    NO_SCALE, SCALE_TO_FILL, SCALE_TO_FIT
+}
